move index.js to app.jsx and add new index.js shim
Note that the code inside app.jsx is exactly what used to be in index.js
so this change causes no functional difference and just represents
preparation for upcoming changes. It simplifies at least two upcoming
changes: one wants the app code to be in a different file and the other
is simplified by JSX syntax.