detect project save based on content type
Also, if saving a file is canceled, only try to cancel project save
telemetry if the save was a project save.

Also also, don't skip showing the save dialog just because a file
doesn't have an extension. Instead, just skip trying to determine
filters for the extension in this case.