add script to merge scratch-gui deps
This is in preparation for having webpack build scratch-gui source
instead of consuming its build output. This script merges the
dependencies listed in package-deps.json with those listed in
scratch-gui's package.json, using the package-deps.json version in case
of conflict, then writes the result into the `devDependencies` section
of scratch-desktop's package.json.